--- app.js
@@ -1,28 +1,0 @@
-require("dotenv").config();
-const express = require("express");
-const app = express();
-const bodyParser = require("body-parser");
-const cookieParser = require("cookie-parser");
-const cors = require("cors");
-
-//Routes
-const indexRoutes = require("./routes/index");
-
-//Middlewares
-app.use(bodyParser.json());
-app.use(cookieParser());
-app.use(cors());
-
-//My Routes
-app.use("/api", indexRoutes);
-
-//PORT
-const port = process.env.PORT || 8000;
-
-app.get("/", (req, res) => {
-  res.send("Server is running.");
-});
-
-app.listen(port, () => {
-  console.log(`app is running at ${port}`);
-});
