+++ package.json
@@ -1,5 +1,5 @@
 {
-  "name": "test1",
+  "name": "fmdapi-node-weaver",
   "version": "1.0.0",
   "description": "",
   "main": "index.js",
